bugfix/minorfix
1.       Delete the start dates for abstract and presentation
submissions, and
2.       Please delete the due date for presentation submission
3.       Please add the word “Organizing” in front of the “Committee”
as we now have two committees
Please let me know when this is done.
4.	  the B.E.S.T. dots are still there on top of the conference home
page and in the main title. In addition, when you take all the dots
out, like you have done in the explanation further down on the page,
please put the letters together like BEST, do not leave them separated
like this: B E S T